Updates build workflow to exclude exclude dev/test dependencies.
- For development, continue to `bundle install` to update gem dependencies.
- For building a release, continue to `./vendor_gems` to update gem
  dependencies.

However:

- `vendor/cache` is now git-ignored, as it is not used in the packaging
  of the release. `Gemfile.lock` will continue to ensure developers
  share the same gem versions.
- The CPI spec, packaging script, and `vendor_gems` script are updated
  to package non-development gems to `vendor/package` and, subsequently,
  include only those in the packaging for the release.

Note that we specifically choose `vendor/package` as the custom gem
packaging path, because `vendor/cache` is a special path recognized by
Bundler and the simple presence of that path means that Bundler will
update gems there. This is fine for development, but we need to be more
selective for the "production" gem set.