CPI is no longer built as a gem.
The BOSH external CPIs do not need to be packaged as gems, so the
gemspec has been remove and all dependency definitions moved to
`Gemfile`. Rubygems updates the `LOAD_PATH` to add <project_dir>/lib`. So,
this change means that "entry points" (bin files and `spec_helper.rb`) need
to perform that modification for the CPI implementation to be found.