Change service principal bash script link
Currently the link will direct you to the github page of the file. This causes a minor annoyance to people wishing to download the script directly from the link or piping into bash on their local machines.

Changing the relative link to the raw script will resolve this issue.

redirecting the powershell script to the raw link as well