Create and delete public IPs on VMs when desired
Introduces new resource_pool/vm_type option called :assign_dynamic_public_ips, which defaults to false.  When set to true, the CPI will create a matching public IP when the NIC is created for a VM.  It will also destroy any matching public IPs during the VM delete phase, regardless of whether the option is enabled or not.