Public IP addresses created on VMs in AZs will use Standard SKU
AZs require IPs to use Standard SKU, but the operators don't tell the
CPI what SKU to use. As a simple workaround, we can see that a VM is
going to placed in a AZ, and therefore you'd have to use Standard SKU
IPs. There may be other times you'd want/need to use the new IP
offering, but this fix only targets this one case.

Signed-off-by: David Stevenson <dstevenson@pivotal.io>