It feels like bundler is broken here
Could not get it to vendor the linux copy of the nokogiri gem
via the script. Did manage to get it to vendor all 3 copies, but then
it also vendored all the dev/test gems too for some reason.

Manually removed those, and added the other platform gems for nokogiri