Attempt to fix failing CI test
The spec/integration/lifecycle_spec.rb:43 integration test has
consistently been failing on CI, giving us an error that suggests the
network interface to which the public ip we're trying to delete hasn't
yet been removed. When re-running this test manually on a hijacked
Concourse container it does not seem to fail. Our current hypothesis is
that the delete occurs too soon after create and the network
interface hasn't yet been created or for whatever reason cannot be
found. We're going to try to sleep to make sure there isn't a race
condition occuring.

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>