Fixed the `Bosh::AzureCloud::Cloud.initialize` method's signature so that the 2nd argument is optional.
Note: Based on the changes made in commit #6c6ec5909b3635d5ea14a5f4dfa2263f2edf1468 (which added the 2nd argument), it seems that the argument was intended to be optional from the start. However, according to the method's signature, it was technically a required argument (despite that the method handles nil/missing values internally on line 29). Also, the following code already doesn't pass the 2nd argument:
- https://github.com/cloudfoundry/bosh-azure-cpi-release/blob/9754cf84f18be7ce0250518bd5a660a459470b56/src/bosh_azure_cpi/bin/bosh_azure_console#L29

Note: The separate handling of the argument's default value on lines 26 and 29 are not redundant. The default value in the signature (line 26) only applies when the optional arg is omitted by the caller. If the caller explicitly specifies the argument with a value of `nil`, the arg default doesn't override the nil, but the `|| 1` on line 29 will. And without the argument default on line 26, the argument technically isn't optional, it is required. Meaning that it is technically an error for any caller to omit that argument, and that line 29 would therefore be covering up the caller's misuse of the method. So both "defaults" are needed in order to correctly handle all of the possible method invocation options for that argument.

Here is the error message which was previously being detected (prior to this change):
> Missing argument 'api_version'. Required 2, 'options, api_version'