🐞 Integration tests pass in CPI version (2)
The latest commits check to make sure that the `api_version` was passed
in to the CPI along with the other arguments (`method`, `arguments`,
`context`); however, our integration tests were not updated to pass in
the API version, and thus our integration tests failed.

Our integration tests now pass in the API version (2, in case you're
interested).

fixes:
```
expected: nil
     got: {"type"=>"Bosh::Clouds::CloudError", "message"=>"Invalid api_version ''", "ok_to_retry"=>false}
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>