Fix gem vendoring during BOSH packaging
The Ruby 3.1 upgrade seemed to no longer be respecting the
BUNDLE_CACHE_PATH environment variable. This caused the packaging script
for the CPI to install gems directly from rubygems.org, which broke in
customer environments that don't have internet access.

We also discovered a similar problem where the Ruby 3.1 package uses
Bundler 2.3.5, but the Gemfile.lock requires 2.3.6, so it tried to
install that from the internet too.

The main issue seems to have been that BUNDLE_DEPLOYMENT was not set
during the `bundle exec` call, which caused bundler to look in the wrong
place. Bundler is in the process of deprecating the `--deployment` flag
on bundle install, because previously it had persisted its value, which
is no longer true. So now we explicitly set the environment variable
where needed.

Signed-off-by: Brian Upton <bupton@vmware.com>