set ruby version for tasks from cpi-certification
the image of cpi-certification is trying to change ruby
to 2.4.4 if chruby is available.

the remote tasks definitions were updated to suppoer the
RUBY_VERSION param to be able to override it if the image
itself has been overriden. The remote files are used in
multiple pipelines and therefore changing the default could've
broken CI in other places.