Fix broken offline environment create-env
bundler stopped picking up some cli flags and some env vars. there have
been warnings around this but we do not see them when the create-env
succeeds. It wasn't caught in CI since the CI environment is not
internetless. So bundler was still able to reach out to get the
gems that were already present locally (which it ignored because it
ignored the cli flags)

Signed-off-by: jpalermo <jpalermo@vmware.com>