Remove config disk support
Originally added in response to issue #472, which doesn't have a lot of details, but does mention
Virtual Machine Scale Sets. This doesn't seem to be a fully implemented feature and there is no
recent progress on making a real feature.

The config disk code is current breaking due to new api changes introduce as part of PR #674
Rather than fix those failures, it makes more sense to remove an unused feature.