Updates API version in most regions
This commit:
* Should make the CPI work correctly in the China, German, and US
  Government regions.
* Moves the API versions out to a common place to avoid future regional
  API version mismatch.
* Updates documentation in comments to point to now-currently-used API
  versions.
* Removes documentation in comments that didn't seem to point to
  anything relevant.

A previous commit updated the Compute API version for the "Generic"
region to account for changes made in a merge PR. However, the API
version was not updated in the other regions. This caused failures for a
customer operating in the China region, and will likely cause failures
in the US Government and German regions.

To avoid this problem in the future, we've moved out the duplicated data
into a single place.

Signed-off-by: Aram Price <pricear@vmware.com>