Attempt to resolve gem issue in CI
When running BATS in CI, the initial deploy of the director succeeds in
one task, but when trying to run bosh delete-env in a teardown task, the
CPI is unable to find the gems it needs. After investigating, it seems
like the gem installation is occurring at a global level instead of
relative to the Ruby package in the CPI.

This commit alters the packaging script to mirror how the AWS CPI and
BOSH Director handle bundle configuration and gem install. The key
difference appears to be that the bosh_bundle_local command runs:
```
bundle config set --local path "${BOSH_INSTALL_TARGET}/gem_home"
```
which explicitly sets the gem installation directory.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>