Treat all resource groups as optional
This commit makes it so that if jq is unable to find a resource group
name of interest in the environment metadata file, it does NOT fail the
cleanup script, but rather just deletes what it can find.

The idea here is that the script iterates over all resource group names
that it knows about, and if it doesn'tt know about some of them, that's
fine... we can assume that Terraform already deleted them... or maybe
they never existed in the first place.

This is important because SOMETIMES, Terraform fails to clean up its
Terraform Workspace because it erroneously thinks that resources that it
successfully deleted have not be actually deleted. (See output from CI
at the end of this message).

Given that every IAAS resource that Terraform managed is actually
deleted, and that this "workspace" is a local concept, and that the
Terraform state files are never going to leave the Concourse Worker, it
seems safe to ignore weirdness caused by this Terraform failure mode.

We theorize that the ensure-cleanup script is failing because the
Terraform resource WOULD have removed all of the resource group names
from the 'environment/metadata' file, had it actually succeeded.

=====================================================================

Destroy complete! Resources: 11 destroyed.
▲ ▲ ▲ ▲ ▲ ▲ ▲ ▲ ▲ ▲ Terraform Destroy ▲ ▲ ▲ ▲ ▲ ▲ ▲ ▲ ▲ ▲
2024/11/06 23:27:59 Error running `workspace delete`: exit status 1, Output: ╷
│ Error: Workspace is not empty
│
│ Workspace "spectrum-wolf" is currently tracking the following resource
│ instances:
│   - azurerm_storage_account.azure_bosh_sa
│   - azurerm_public_ip.azure_ip_bosh
│   - azurerm_resource_group.azure_rg_bosh
│   - azurerm_storage_container.azure_stemcell_container
│   - azurerm_subnet.azure_bats_subnet
│   - azurerm_storage_container.azure_bosh_container
│   - azurerm_subnet.azure_bats_subnet_2
│   - azurerm_subnet.azure_bosh_subnet
│   - azurerm_network_security_group.azure_bosh_nsg
│   - azurerm_public_ip.azure_ip_bats
│   - azurerm_virtual_network.azure_bosh_network
│
│ Deleting this workspace would cause Terraform to lose track of any
│ associated remote objects, which would then require you to delete them
│ manually outside of Terraform. You should destroy these objects with
│ Terraform before deleting the workspace.
│
│ If you want to delete this workspace anyway, and have Terraform forget
│ about these managed objects, use the -force option to disable this safety
│ check.
╵

Signed-off-by: Ming Xiao <ming.xiao@broadcom.com>