Do not try to create availability set with more fault domains than the region supports when migrating from unmanaged to managed disks
Part of the migration is update the availability set and just using the existing fault domains on the old set. In unmanaged
availability sets, it's possible to have more fault domains than the region supports but this is not supported for managed
availability sets so it throws an error.

Now we use the lesser of the region maximum and the previous setting. This does involve fetching the maximum from the region
for each set migration, which is a large API payload from Azure, but once a deployment has been migrated it won't
need to make these calls again.