Remove integration specs for unmanaged disks
Unmanaged disks in Azure will be deprecated in September 2025. Starting
with the release of Windows stemcell 2019.84 earlier this year, new
images and marketplace offerings can no longer be used to launch VMs
using unmanaged disks. Attempts to do so result in the following error:
```
The specified platform image is not supported for creating a Virtual
Machine with unmanaged disks. Please refer to the disallowed VM disk
types at
https://docs.microsoft.com/en-us/rest/api/compute/virtualmachineimages/get#vmdisktypes.
```

Rather than pin back to an obsolete stemcell in CI, this removes the
integration tests related to unmanaged disks. This commit purposefully
leaves unit tests and related code around unmanaged disks; these should
be removed closer to the final September deadline for migration.